refactor: move Block{Request, Response} to node-network
