--- node/router/messages/src/block_request.rs
@@ -1,31 +1,0 @@
-// Copyright (c) 2019-2026 Provable Inc.
-// This file is part of the snarkOS library.
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at:
-
-// http://www.apache.org/licenses/LICENSE-2.0
-
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use super::*;
-
-use snarkos_node_bft_events::EventTrait;
-
-use std::borrow::Cow;
-
-/// Re-export the BlockRequest structure from BFT.
-pub use snarkos_node_bft_events::BlockRequest;
-
-impl MessageTrait for BlockRequest {
-    /// Returns the message name.
-    #[inline]
-    fn name(&self) -> Cow<'static, str> {
-        EventTrait::name(self)
-    }
-}
